refactor: split frame.h into frame.h and frame.cc
